+++ tests/Traits/PdfRoundedRectangleTraitTest.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace fpdf\Tests\Traits;
 
-use fpdf\Color\PdfRgbColor;
 use fpdf\Enums\PdfMove;
 use fpdf\PdfException;
 use fpdf\PdfRectangle;
@@ -24,8 +23,7 @@ final class PdfRoundedRectangleTraitTest extends TestCase
 {
     public function testMoveBelow(): void
     {
-        $document = new PdfDocumentRoundedRectangle();
-        $document->addPage();
+        $document = $this->createDocument();
         $document->roundedRect(100, 50, 5, 10, 20, move: PdfMove::BELOW);
         self::assertSame(10.0, $document->getX());
         self::assertSame(70.0, $document->getY());
@@ -33,8 +31,7 @@ final class PdfRoundedRectangleTraitTest extends TestCase
 
     public function testMoveNewLine(): void
     {
-        $document = new PdfDocumentRoundedRectangle();
-        $document->addPage();
+        $document = $this->createDocument();
         $document->roundedRect(100, 50, 5, 10, 20, move: PdfMove::NEW_LINE);
         self::assertSame($document->getLeftMargin(), $document->getX());
         self::assertSame(70.0, $document->getY());
@@ -42,65 +39,31 @@ final class PdfRoundedRectangleTraitTest extends TestCase
 
     public function testMoveRight(): void
     {
-        $document = new PdfDocumentRoundedRectangle();
-        $document->addPage();
+        $document = $this->createDocument();
         $document->roundedRect(100, 50, 5, 10, 20);
         self::assertSame(110.0, $document->getX());
         self::assertSame(20.0, $document->getY());
     }
 
-    public function testRoundedInvalidRadius(): void
-    {
-        $document = new PdfDocumentRoundedRectangle();
-        self::expectException(PdfException::class);
-        self::expectExceptionMessage('Invalid radius: 40, maximum allowed: 25.');
-        $document->roundedRect(50, 100, 40);
-    }
-
     public function testRoundedRadiusNotPositive(): void
     {
-        $document = new PdfDocumentRoundedRectangle();
+        $document = $this->createDocument();
         self::expectException(PdfException::class);
         self::expectExceptionMessage('The radius must be positive, 0 given.');
         $document->roundedRect(50, 100, 0);
     }
 
-    public function testRoundedRect(): void
-    {
-        $document = new PdfDocumentRoundedRectangle();
-        $document->addPage();
-        $document->setFillColor(PdfRgbColor::darkGray());
-        $document->setDrawColor(PdfRgbColor::red());
-
-        $document->setXY(20, 20);
-        $document->setLineWidth(1.5);
-        $document->roundedRect(
-            width: 20,
-            height: 20,
-            radius: 5,
-            move: PdfMove::BELOW
-        );
-        self::assertSame(20.0, $document->getX());
-        self::assertSame(40.0, $document->getY());
-
-        $document->setLineWidth(0.5);
-        $document->roundedRect(
-            width: 100,
-            height: 10,
-            radius: 5,
-            move: PdfMove::BELOW
-        );
-        self::assertSame(20.0, $document->getX());
-        self::assertSame(50.0, $document->getY());
-    }
-
     public function testRoundedRectangle(): void
     {
         $rect = new PdfRectangle(10, 20, 100, 50);
-        $document = new PdfDocumentRoundedRectangle();
-        $document->addPage();
+        $document = $this->createDocument();
         $document->roundedRectangle($rect, 5);
         self::assertSame($rect->right(), $document->getX());
         self::assertSame($rect->y, $document->getY());
+    }
+
+    private function createDocument(): PdfDocumentRoundedRectangle
+    {
+        return (new PdfDocumentRoundedRectangle())->addPage();
     }
 }